+++ tangent-cc-lib/src/lib/model/key-code.models.ts
@@ -97,6 +97,7 @@ export type NonWSKCode =
   | 'F23'
   | 'F24'
   | 'PrintScreen'
+  | 'ScrollLock'
   | 'Delete'
   | 'ArrowDown'
   | 'ArrowLeft'