+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- add scroll lock action
+
 ## [0.0.53](https://github.com/andy23512/tangent-cc-lib/releases/tag/0.0.53)
 
 - convertKeyboardLayoutToCharacterKeyCodeMap keep all mapping to a character